Ignore OMEMO messsage in MUC if sender is not present.
This can happen if there is an additional <body> element,
since the messages are then saved and replayed on entering as chat history.
Should help with #65.

As the message is then completely ignored by this plugin,
This also means that the text in the <body> will be displayed to the user.