Improve handling of messages to own account.
The fix in the last commit uncovered that such messages do not
always have a "to" attribute (at least they do not with Prosody).
"Uncovered" means this was completely broken.

Additionally, lurch now ignores a decryption failure under specific
circumstances: If both sender and recipient are the own account,
and the error code is libsignal's "duplicate message", a warning
is written to the debug log and the message itself ignored, as it
was probably a carbon-copy.